requirements.txt
Django>=6.0.0
numpy>=2.0.0
opencv-python>=4.10.0
pillow>=10.0.0
face-recognition>=1.3.0
dlib-bin>=20.0.0
setuptools<70
django-crispy-forms
crispy-bootstrap5
django-widget-tweaks
openpyxl
requests
reportlab
python-dateutil
